feat: rename cookies folder to useCookieListener and support className/style for all components
