+++ src/index.ts
@@ -3,7 +3,6 @@ export { default as Catalogue } from './catalogue';
 export { default as CollapsibleActionItems } from './collapsibleActionItems';
 export { default as ContentLayout } from './contentLayout';
 export { default as ContextMenu } from './contextMenu';
-export { default as useCookieListener } from './cookies';
 export { default as Copy } from './copy';
 export { default as Drawer } from './drawer';
 export { default as Dropdown } from './dropdown';
@@ -28,6 +27,7 @@ export { default as SpreadSheet } from './spreadSheet';
 export { default as StatusTag } from './statusTag';
 export { default as Table } from './table';
 export { default as TinyTag } from './tinyTag';
+export { default as useCookieListener } from './useCookieListener';
 export { default as useDebounce } from './useDebounce';
 export { default as useIntersectionObserver } from './useIntersectionObserver';
 export { default as useList } from './useList';
